requirements.txt
asgiref==3.4.1
Deprecated==1.2.13
Django==3.2.9
django-cleanup==5.2.0
django-debug-toolbar==3.2.2
django-phonenumber-field==6.0.0
django-redis==5.2.0
djangorestframework==3.12.4
mysqlclient==2.0.3
packaging==21.3
phonenumberslite==8.12.40
Pillow==8.4.0
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
sqlparse==0.4.2
wrapt==1.13.3
